rss-intel/backend/pyproject.toml
[project]
name = "rss-intel-backend"
version = "1.0.0"
description = "RSS Intelligence Dashboard Backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "alembic>=1.12.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "httpx>=0.27.0",
    "apscheduler>=3.10.0",
    "pyyaml>=6.0",
    "python-multipart>=0.0.6",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "Pillow>=10.0.0",
    "blurhash>=1.1.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "python-dateutil>=2.9.0",
    "feedparser>=6.0.0",
    "numpy>=1.21.0",
    "newspaper3k>=0.2.8",
    "readability-lxml>=0.8.1",
    "aiohttp>=3.9.0",
    "asyncio-throttle>=1.0.2",
    "pgvector>=0.2.5",
    "scikit-learn>=1.5",
    "pandas>=2.2",
    "numpy>=1.26",
    "sentence-transformers>=3.0",
    "torch>=2.2",
    "orjson>=3.9",
    "weaviate-client>=4.6",
    "rank-bm25>=0.2",
    "rapidfuzz>=3.9",
    "playwright>=1.40.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
asyncio_mode = "auto"